python/requirements.txt
pwntools
lief
pygments
graphviz
